+++ Abascus/DLCCraft/common/State.java
@@ -2,9 +2,9 @@ package Abascus.DLCCraft.common;
 
 public class State 
 {
-	public byte state = 0;
+	public int state = 0;
 	
-	public State(byte i)
+	public State(int i)
 	{
 		state = i;
 	}
